refactor: jobManager
